Version 1.0
The program works in this state and can accept as input multiple LS-Dyna keyfiles and produce the node and mesh outputs. The original keyword information is lost, so there is no way to tell whether, for example, a four-node mesh connectivity refers to a tet solid element or a quad surface element.